chore: export run in main.js
